+++ R/CoreSet-class.R
@@ -1,5 +1,6 @@
 #' @include CoreSet-class.R LongTable-class.R
 #' @importClassesFrom MultiAssayExperiment MultiAssayExperiment
+#' @importFrom MultiAssayExperiment MultiAssayExperiment
 #' @import checkmate
 NULL